extract multi modal agent loop test to a new file
